refactor(frontend): split TodoList into sub-components
Split TodoList (720→1 line) into:
- todo-list/SkeletonRow.tsx: skeleton loading row
- todo-list/TodoItemRow.tsx: todo list item row
- todo-list/index.tsx: main component

All exports maintained via re-export from original file.
