+++ frontend/src/components/TodoList.tsx
@@ -1,720 +1,6 @@
-import { useState, useEffect, useMemo, useCallback } from 'react';
-import { useApp } from '@/hooks/useApp';
-import { Empty, Input, Skeleton, Checkbox, Modal, App as AntApp } from 'antd';
-import { ClockCircleOutlined, InboxOutlined, SearchOutlined, SwapOutlined, StopOutlined, CopyOutlined, DragOutlined } from '@ant-design/icons';
-import { StatusPicker } from './StatusPicker';
-import * as db from '@/utils/database';
-import type { ProjectDirectory, Todo } from '@/types';
-import { ExecutorBadge } from './ExecutorBadge';
-import { LoopListPanel } from './LoopStudioListPanel';
-import type { LoopListItem } from '@/types/loop';
-import * as dbLoops from '@/utils/database/loops';
-import { EXECUTORS_FOR_PICKER } from '@/types/execution';
-import { ExecutorPicker } from './todo-drawer/ExecutorPicker';
-import { ActionToolbar, type BatchActionItem } from './common/ActionToolbar';
-import { formatRelativeTime } from '@/utils/datetime';
-import { WorkspaceSelect } from './common/WorkspaceSelect';
+// TodoList 重新导出入口
+//
+// 本文件已重构，所有子组件拆分到 todo-list/ 目录。
+// 为保持 API 兼容，继续从此文件重新导出。
 
-interface TodoListProps {
-  onOpenCreateModal: () => void;
-  onSelectTodo?: (todoId: string | number) => void;
-  onSelectLoop?: (loopId: number) => void;
-  onCreateLoop?: () => void;
-  loopUpdateCount?: number;
-  forcedListMode?: 'item' | 'loop';
-  onListModeChange?: (mode: 'item' | 'loop') => void;
-  hideCreateButton?: boolean;
-}
-
-function SkeletonRow() {
-  return <div className="skeleton-row" />;
-}
-
-function SkeletonList() {
-  return (
-    <div style={{ padding: '12px 16px' }}>
-      {Array.from({ length: 6 }).map((_, i) => (
-        <SkeletonRow key={i} />
-      ))}
-    </div>
-  );
-}
-
-export function TodoList(props: TodoListProps) {
-  const { onOpenCreateModal, onSelectTodo, onSelectLoop, onCreateLoop, loopUpdateCount, forcedListMode, onListModeChange, hideCreateButton } = props;
-  const { state, dispatch } = useApp();
-  const { todos, selectedTodoId, selectedTagId, selectedWorkspace, tags } = state;
-  const { message } = AntApp.useApp();
-  const [isLoading, setIsLoading] = useState(true);
-  // 搜索关键字状态，用于按标题或提示词过滤 todo 列表
-  const [searchKeyword, setSearchKeyword] = useState('');
-  // 列表模式：'item' = 事项, 'loop' = 环路
-  const [listMode, setListMode] = useState<'item' | 'loop'>(() => {
-    const saved = localStorage.getItem('ntd_list_mode');
-    if (saved === 'item' || saved === 'loop') return saved;
-    return 'item';
-  });
-  // 环路列表数据（只在 listMode === 'loop' 时使用）
-  const [loopList, setLoopList] = useState<LoopListItem[]>([]);
-  const [loopLoading, setLoopLoading] = useState(false);
-  // 当前选中的 loop id（来自左侧环路列表），用于高亮
-  const [selectedLoopId, setSelectedLoopId] = useState<number | null>(null);
-  // 项目目录：工作空间选择器需要目录列表
-  const [projectDirectories, setProjectDirectories] = useState<ProjectDirectory[]>([]);
-  // 标记项目目录初始加载是否已完成（无论成功与否），用于守卫环路加载等工作空间确定后再发起请求
-  const [directoriesReady, setDirectoriesReady] = useState(false);
-  // —— 通用工具栏：跨模式的多选 id 列表 ——
-  // 切换 listMode 时清空，避免不同模式 id 串台（todo/loop 都是 number id）
-  const [selectedIds, setSelectedIds] = useState<number[]>([]);
-  // 批量更换执行器 Modal（事项模式）
-  const [executorModalOpen, setExecutorModalOpen] = useState(false);
-  const [pendingExecutorChangeIds, setPendingExecutorChangeIds] = useState<number[]>([]);
-  // 强停确认 Modal（环路）
-  const [forceStopModalOpen, setForceStopModalOpen] = useState(false);
-  const [pendingForceStopIds, setPendingForceStopIds] = useState<number[]>([]);
-  // 批量工作空间操作 Modal（事项/环路共用）
-  const [workspaceBatchModalOpen, setWorkspaceBatchModalOpen] = useState(false);
-  const [workspaceBatchMode, setWorkspaceBatchMode] = useState<'copy' | 'move'>('copy');
-  // target 改为工作空间 id（project_directories.id 唯一键），不再用 path。
-  const [workspaceBatchTarget, setWorkspaceBatchTarget] = useState<number | null>(null);
-  const [pendingWorkspaceBatchIds, setPendingWorkspaceBatchIds] = useState<number[]>([]);
-  const [workspaceBatchProcessing, setWorkspaceBatchProcessing] = useState(false);
-  const [workspaceBatchContext, setWorkspaceBatchContext] = useState<'item' | 'loop'>('item');
-
-  useEffect(() => {
-    setIsLoading(false);
-  }, []);
-
-  // 外部导航强制指定 listMode 时（例如左侧主导航点击“事项/环路”），
-  // 需要让内部状态跟随，否则中间列表会停留在用户上次切换的模式。
-  useEffect(() => {
-    if (!forcedListMode) return;
-    if (forcedListMode === listMode) return;
-    setListMode(forcedListMode);
-  }, [forcedListMode, listMode]);
-
-  // 进入页面时拉取项目目录；后续 Todo 抽屉新增/删除目录时也会主动重拉，确保分组始终准确。
-  // 失败时静默处理：分组视图退化为只显示路径即可，不阻塞主流程。
-  const reloadProjectDirectories = useCallback(() => {
-    db.getProjectDirectories() // 从后端拉取全量目录列表
-      .then(dirs => {
-        setProjectDirectories(dirs);
-        setDirectoriesReady(true);
-      })
-      .catch(() => {
-        setDirectoriesReady(true); // 失败也标记为 ready，避免阻塞环路加载
-      });
-  }, []);
-
-  // 首次加载目录后，如果没有选中工作空间且存在目录，自动选中第一个。
-  // 这样保证用户必须工作在某个空间下，实现工作空间隔离。
-  useEffect(() => {
-    reloadProjectDirectories(); // 首次加载目录
-    // 监听 TodoDrawer 快速新增项目目录的事件，及时刷新分组数据
-    const handleDirAdded = () => reloadProjectDirectories();
-    window.addEventListener('projectDirectoryAdded', handleDirAdded); // 跨组件通知
-    return () => window.removeEventListener('projectDirectoryAdded', handleDirAdded); // 清理：卸载时移除监听
-  }, [reloadProjectDirectories]);
-
-  // 当目录列表加载完成后，若当前未选中任何工作空间且存在目录，自动选中第一个。
-  // 自动选中用 id 而非 path —— selectedWorkspace 现已统一为 project_directories.id。
-  useEffect(() => {
-    if (projectDirectories.length > 0 && selectedWorkspace === null) {
-      dispatch({ type: 'SELECT_WORKSPACE', payload: projectDirectories[0].id });
-    }
-  }, [projectDirectories, selectedWorkspace, dispatch]);
-
-  // 当列表切换到「环路」时，自动加载 loop 列表；或环路变更时刷新
-  // 必须等待项目目录加载完成且工作空间已确定，否则请求不携带 workspace 参数返回错误数据。
-  // selectedWorkspace 现在就是 id（唯一键），直接传给 listLoops 即可，不需要再做 path→id 推导。
-  useEffect(() => {
-    if (listMode !== 'loop') return;
-    if (!directoriesReady) return;
-    // 已有目录但尚未选定工作空间 → 等待 auto-select 后再加载
-    if (projectDirectories.length > 0 && selectedWorkspace === null) return;
-    setLoopLoading(true);
-    dbLoops.listLoops(selectedWorkspace)
-      .then(setLoopList)
-      .catch(() => setLoopList([]))
-      .finally(() => setLoopLoading(false));
-  }, [listMode, loopUpdateCount, selectedWorkspace, projectDirectories, directoriesReady]);
-
-  // 切换工作空间后按需拉取该工作空间的 todo 列表。
-  // 分桶改造后，切换 workspace 时如果目标桶为空（从未加载过），
-  // 必须主动拉一次才能让 useVisibleTodos() 返回数据。
-  // 监听 selectedWorkspace + directoriesReady 确保目录已加载完毕后
-  // 才发请求；不监听 todos 长度避免无限循环。
-  useEffect(() => {
-    if (!directoriesReady) return;
-    if (selectedWorkspace == null) return;
-    // 仅在 item 模式下触发 todo 拉取（loop 模式走上方 listLoops effect）
-    if (listMode !== 'item') return;
-    // 桶已有数据时跳过——已在 store 中，不会因切换 workspace 而丢失。
-    // 兜底：如果桶确实为空，也会拉一次（空 workspace 拉回空列表，只一次）。
-    const currentBucket = state.todosByWorkspace?.[selectedWorkspace];
-    if (currentBucket !== undefined) return;
-    db.getAllTodos(selectedWorkspace).then(todos => {
-      dispatch({ type: 'SET_TODOS_BY_WORKSPACE', workspaceId: selectedWorkspace, payload: todos });
-    });
-  }, [selectedWorkspace, directoriesReady, listMode, state.todosByWorkspace, dispatch]);
-
-  // 持久化列表模式到 localStorage
-  useEffect(() => {
-    localStorage.setItem('ntd_list_mode', listMode);
-  }, [listMode]);
-
-  // 向壳层同步当前列表模式，便于左侧主导航高亮与全局路由状态保持一致。
-  useEffect(() => {
-    onListModeChange?.(listMode);
-  }, [listMode, onListModeChange]);
-
-  // 切换 listMode 时清空选择：todo/loop 虽然 id 都是 number，
-  // 但语义不同（同一数字可能指向不同实体），跨模式保留选择会让用户困惑。
-  useEffect(() => {
-    setSelectedIds([]);
-    setSelectedLoopId(null);
-  }, [listMode]);
-
-  // 切换单条 id 的选中态（toggle 语义，工具栏的「全选」用 onSelectionChange 全量覆盖）
-  const toggleSelect = useCallback((id: number) => {
-    setSelectedIds(prev => prev.includes(id) ? prev.filter(x => x !== id) : [...prev, id]);
-  }, []);
-
-  const filteredTodos = useMemo(() => {
-    // 环路模式下不需要过滤 todo（左侧渲染环路列表）
-    if (listMode === 'loop') return [];
-
-    // 先按标签过滤
-    // 按选中标签过滤：直接读 Todo.tag_ids 即可，
-    // 不需要 `as any` — Todo 类型已在 frontend/src/types/todo.ts 中声明该字段。
-    // 显式用 `!== null` 判定而不是真值判断：selectedTagId 类型是 number | null，
-    // 0 是合法 id，truthy 判定会把合法的 0 当作「未选中」而错误地跳过过滤。
-    let result = selectedTagId !== null
-      ? todos.filter(t => t.tag_ids?.includes(selectedTagId))
-      : todos;
-    
-    // 再按关键字搜索（匹配标题或提示词）
-    if (searchKeyword.trim()) {
-      const keyword = searchKeyword.toLowerCase().trim();
-      result = result.filter(todo => {
-        const title = (todo.title || '').toLowerCase();
-        const prompt = (todo.prompt || '').toLowerCase();
-        return title.includes(keyword) || prompt.includes(keyword);
-      });
-    }
-
-    return result;
-  }, [todos, selectedTagId, searchKeyword, listMode]);
-
-  const filteredLoopList = useMemo(() => {
-    const keyword = searchKeyword.trim().toLowerCase();
-    if (!keyword) return loopList;
-    return loopList.filter(l => (l.name || '').toLowerCase().includes(keyword));
-  }, [loopList, searchKeyword]);
-
-  // 当前 listMode 下"可见可选"的 id 列表，传给 ActionToolbar 用于「全选」/计数。
-  // 两种模式都按当前 searchKeyword 过滤后的列表计算，避免「全选」选中隐藏项。
-  const visibleIds = useMemo<number[]>(() => {
-    if (listMode === 'item') return filteredTodos.map(t => t.id);
-    return filteredLoopList.map(l => l.id);
-  }, [listMode, filteredTodos, filteredLoopList]);
-
-  const handleStatusChange = useCallback(async (todoId: number, title: string, prompt: string, newStatus: string) => {
-    try {
-      const updated = await db.updateTodo(todoId, title, prompt, newStatus);
-      dispatch({ type: 'UPDATE_TODO', payload: updated });
-    } catch {
-      // ignore: interceptor already shows error
-    }
-  }, [dispatch]);
-
-  // —— 批量操作：事项模式 ——
-  // 「更换执行器」打开 Modal，Modal 内确认后调 db.batchUpdateTodosExecutor
-  const openItemChangeExecutor = useCallback((ids: number[]) => {
-    setPendingExecutorChangeIds(ids);
-    setExecutorModalOpen(true);
-  }, []);
-
-  // —— 批量操作：环路模式 ——
-  const openLoopForceStop = useCallback((ids: number[]) => {
-    setPendingForceStopIds(ids);
-    setForceStopModalOpen(true);
-  }, []);
-
-  // —— 批量操作：工作空间复制/移动（事项模式） ——
-  const openItemCopyWorkspace = useCallback((ids: number[]) => {
-    setWorkspaceBatchContext('item');
-    setWorkspaceBatchMode('copy');
-    setPendingWorkspaceBatchIds(ids);
-    setWorkspaceBatchTarget(null);
-    setWorkspaceBatchModalOpen(true);
-  }, []);
-  const openItemMoveWorkspace = useCallback((ids: number[]) => {
-    setWorkspaceBatchContext('item');
-    setWorkspaceBatchMode('move');
-    setPendingWorkspaceBatchIds(ids);
-    setWorkspaceBatchTarget(null);
-    setWorkspaceBatchModalOpen(true);
-  }, []);
-
-  // —— 批量操作：工作空间复制/移动（环路模式） ——
-  const openLoopCopyWorkspace = useCallback((ids: number[]) => {
-    setWorkspaceBatchContext('loop');
-    setWorkspaceBatchMode('copy');
-    setPendingWorkspaceBatchIds(ids);
-    setWorkspaceBatchTarget(null);
-    setWorkspaceBatchModalOpen(true);
-  }, []);
-  const openLoopMoveWorkspace = useCallback((ids: number[]) => {
-    setWorkspaceBatchContext('loop');
-    setWorkspaceBatchMode('move');
-    setPendingWorkspaceBatchIds(ids);
-    setWorkspaceBatchTarget(null);
-    setWorkspaceBatchModalOpen(true);
-  }, []);
-
-  // 确认工作空间批量操作
-  const handleConfirmWorkspaceBatch = useCallback(async () => {
-    const ids = pendingWorkspaceBatchIds;
-    const target = workspaceBatchTarget;
-    // target 是工作空间 id；null/undefined 都视为未选
-    if (ids.length === 0 || target == null) return;
-    setWorkspaceBatchProcessing(true);
-    try {
-      const mode = workspaceBatchMode;
-      const context = workspaceBatchContext;
-      let result: { updated_count: number; total: number };
-
-      if (context === 'item') {
-        if (mode === 'copy') {
-          result = await db.batchCopyTodosWorkspace(ids, target);
-        } else {
-          result = await db.batchMoveTodosWorkspace(ids, target);
-        }
-      } else {
-        if (mode === 'copy') {
-          result = await dbLoops.batchCopyLoopsWorkspace(ids, target);
-        } else {
-          result = await dbLoops.batchMoveLoopsWorkspace(ids, target);
-        }
-      }
-
-      const actionLabel = mode === 'copy' ? '复制' : '移动';
-      // 提示用工作空间展示名（name 优先、path 兜底），不再直接拼接 id。
-      const targetName = projectDirectories.find(d => d.id === target)?.name
-        ?? projectDirectories.find(d => d.id === target)?.path
-        ?? String(target);
-      if (result.updated_count === result.total) {
-        message.success(`已${actionLabel} ${result.updated_count} 项到「${targetName}」`);
-      } else {
-        message.warning(`${actionLabel}成功 ${result.updated_count} 条，失败 ${result.total - result.updated_count} 条`);
-      }
-
-      // 刷新数据
-      if (context === 'item') {
-        // 批量操作可能跨 workspace（用户从 A 移到 B），但 store 里只会展示当前
-        // workspace 的桶。简化策略：只拉当前 workspace 的桶；若需要跨 workspace
-        // 全量刷新，可单独走 SET_WORKSPACE_TODOS 多次（暂不优化）。
-        const wid = selectedWorkspace;
-        const allItems = wid != null ? await db.getAllTodos(wid) : [];
-        if (wid != null) {
-          dispatch({ type: 'SET_TODOS_BY_WORKSPACE', workspaceId: wid, payload: allItems });
-        }
-      } else {
-        // 环路模式：刷新 loop 列表。selectedWorkspace 现已统一为 id，直接传即可。
-        const loops = await dbLoops.listLoops(selectedWorkspace);
-        setLoopList(loops);
-      }
-
-      setWorkspaceBatchModalOpen(false);
-      setSelectedIds([]);
-    } catch (err) {
-      message.error(`操作失败: ${err instanceof Error ? err.message : '未知错误'}`);
-    } finally {
-      setWorkspaceBatchProcessing(false);
-    }
-  }, [pendingWorkspaceBatchIds, workspaceBatchTarget, workspaceBatchMode, workspaceBatchContext, message, dispatch, selectedWorkspace, projectDirectories]);
-
-  // 确认更换执行器（事项模式）
-  const handleConfirmChangeExecutor = useCallback(async (executor: string) => {
-    const ids = pendingExecutorChangeIds;
-    if (ids.length === 0) return;
-    setExecutorModalOpen(false);
-    setPendingExecutorChangeIds([]);
-    try {
-      const result = await db.batchUpdateTodosExecutor(ids, executor);
-      if (result.failed.length === 0) {
-        message.success(`已为 ${result.updated.length} 项更换执行器为「${executor}」`);
-      } else {
-        message.warning(`成功 ${result.updated.length} 条，失败 ${result.failed.length} 条`);
-      }
-      // 触发列表刷新：item 模式依赖全局 todos 状态，由 useApp 拉取；只拉当前 workspace 桶
-      if (listMode === 'item') {
-        const wid = selectedWorkspace;
-        if (wid != null) {
-          const allItems = await db.getAllTodos(wid);
-          dispatch({ type: 'SET_TODOS_BY_WORKSPACE', workspaceId: wid, payload: allItems });
-        }
-      }
-    } catch {
-      // axios 拦截器已弹错
-    } finally {
-      setSelectedIds([]);
-    }
-  }, [pendingExecutorChangeIds, listMode, message, dispatch]);
-
-  // 确认强停（环路占位实现）
-  const handleConfirmForceStop = useCallback(async () => {
-    const ids = pendingForceStopIds;
-    if (ids.length === 0) return;
-    setForceStopModalOpen(false);
-    setPendingForceStopIds([]);
-    try {
-      const result = await dbLoops.forceStopLoops(ids);
-      if (result.stopped.length > 0) {
-        message.success(`已强停 ${result.stopped.length} 个环路`);
-      } else {
-        // 占位实现会全部走失败分支，统一提示"开发中"
-        message.warning(`环路强停功能开发中（已选 ${ids.length} 个）`);
-      }
-    } finally {
-      setSelectedIds([]);
-    }
-  }, [pendingForceStopIds, message]);
-
-  // 工具栏配置：按 listMode 切换 createLabel / batchActions。
-  // 「新建」按钮统一显示为「新建」2 字（用户能从当前 listMode 知道新建的是什么）；
-  // 批量菜单项按模式差异。
-  const toolbarConfig = useMemo<{
-    createLabel: string;
-    batchActions: BatchActionItem<number>[];
-  }>(() => {
-    if (listMode === 'item') {
-      return {
-        createLabel: '新建',
-        batchActions: [{
-          key: 'change-executor',
-          label: '更换执行器',
-          icon: <SwapOutlined />,
-          onClick: openItemChangeExecutor,
-        }, {
-          key: 'copy-workspace',
-          label: '复制到',
-          icon: <CopyOutlined />,
-          onClick: openItemCopyWorkspace,
-        }, {
-          key: 'move-workspace',
-          label: '移动到',
-          icon: <DragOutlined />,
-          onClick: openItemMoveWorkspace,
-        }],
-      };
-    }
-    return {
-      createLabel: '新建',
-      batchActions: [{
-        key: 'copy-workspace',
-        label: '复制到',
-        icon: <CopyOutlined />,
-        onClick: openLoopCopyWorkspace,
-      }, {
-        key: 'move-workspace',
-        label: '移动到',
-        icon: <DragOutlined />,
-        onClick: openLoopMoveWorkspace,
-      }, {
-        key: 'force-stop',
-        label: '强停',
-        icon: <StopOutlined />,
-        danger: true,
-        onClick: openLoopForceStop,
-      }],
-    };
-  }, [listMode, openItemChangeExecutor, openItemCopyWorkspace, openItemMoveWorkspace, openLoopCopyWorkspace, openLoopMoveWorkspace, openLoopForceStop]);
-
-  const tagMap = useMemo(() => {
-    const map = new Map<number, typeof tags[0]>();
-    for (const tag of tags) map.set(tag.id, tag);
-    return map;
-  }, [tags]);
-
-  // 抽离 Todo 行渲染，平铺与分组两个模式共用，避免重复代码
-  const renderTodoItem = (todo: Todo) => {
-    // tag_ids 在 Todo 类型中是必填 number[]，但历史接口偶发返回缺失字段，
-    // 所以用可选链 + 空数组兜底，避免运行时崩溃。
-    const todoTags = todo.tag_ids?.map(id => tagMap.get(id)).filter((t): t is typeof tags[0] => !!t) ?? [];
-    const primaryTag = todoTags[0];
-    const isCompleted = todo.status === 'completed';
-    const relativeTime = formatRelativeTime(todo.updated_at);
-    const isChecked = selectedIds.includes(todo.id);
-
-    return (
-      <div
-        key={todo.id}
-        onClick={() => {
-          dispatch({ type: 'SELECT_TODO', payload: todo.id });
-          onSelectTodo?.(todo.id);
-        }}
-        className={`todo-item ${selectedTodoId === todo.id ? 'selected' : ''}`}
-        style={{
-          cursor: 'pointer',
-          borderLeftColor: primaryTag?.color || '#cbd5e1',
-          borderLeftWidth: 4,
-          borderLeftStyle: 'solid',
-          // 工具栏的多选复选框用 position: absolute 浮在卡片左上；
-          // 若 .todo-item 不设 position: relative，复选框会逃逸到上层容器，
-          // 所有卡片的复选框都叠在同一个屏幕坐标，点击会命中最后渲染的那个。
-          position: 'relative',
-        }}
-        role="button"
-        tabIndex={0}
-        onKeyDown={(e) => {
-          if (e.key === 'Enter' || e.key === ' ') {
-            e.preventDefault();
-            dispatch({ type: 'SELECT_TODO', payload: todo.id });
-            onSelectTodo?.(todo.id);
-          }
-        }}
-      >
-        {/* 多选复选框：position absolute 浮在卡片左上，避免打乱原本的 layout。
-            stopPropagation 阻止冒泡到卡片的 onClick（不会触发详情选中）。 */}
-        <Checkbox
-          checked={isChecked}
-          onChange={(e) => { e.stopPropagation(); toggleSelect(todo.id); }}
-          onClick={(e) => e.stopPropagation()}
-          data-testid={`todo-row-checkbox-${todo.id}`}
-          style={{ position: 'absolute', top: 12, left: 12, zIndex: 1 }}
-        />
-        <div className="todo-item-content" style={{ paddingLeft: 28 }}>
-          <div className="todo-item-main">
-            <div style={{ display: 'flex', alignItems: 'center', gap: 8, flexWrap: 'wrap' }}>
-              <div
-                className="todo-item-title"
-                style={{ opacity: isCompleted ? 0.6 : 1 }}
-              >
-                <span style={{ color: '#999', marginRight: 4, fontSize: 13 }}>#{todo.id}</span>{todo.title}
-              </div>
-              <ExecutorBadge executor={todo.executor || 'claudecode'} />
-            </div>
-            {todo.prompt && (
-              <div className="todo-item-desc">
-                {todo.prompt.length > 60 ? todo.prompt.substring(0, 60) + '...' : todo.prompt}
-              </div>
-            )}
-            <div className="todo-item-tags" style={{ justifyContent: 'space-between' }}>
-              <div style={{ display: 'flex', alignItems: 'center', gap: 4, flexWrap: 'wrap' }}>
-                {todoTags.map(t => (
-                  <span
-                    key={t.id}
-                    className="todo-tag-badge"
-                    style={{
-                      backgroundColor: t.color + '18',
-                      color: t.color,
-                      border: `1px solid ${t.color}30`,
-                    }}
-                  >
-                    {t.name}
-                  </span>
-                ))}
-                {todo.scheduler_config && (
-                  <ClockCircleOutlined
-                    style={{
-                      fontSize: 12,
-                      color: todo.scheduler_enabled ? 'var(--color-warning)' : 'var(--color-text-tertiary)',
-                      marginLeft: todoTags.length > 0 ? 4 : 0,
-                    }}
-                  />
-                )}
-                {/* todo_type === 1 已废弃：评审模板自 V15 起迁出至 review_templates 表。 */}
-                {todo.todo_type === 2 && (
-                  <span
-                    className="todo-tag-badge"
-                    style={{
-                      backgroundColor: '#13c2c218',
-                      color: '#13c2c2',
-                      border: '1px solid #13c2c230',
-                    }}
-                    title={`评审实例 (原 todo #${todo.parent_todo_id ?? '?'})`}
-                  >
-                    [评审]
-                  </span>
-                )}
-              </div>
-              <span
-                style={{
-                  fontSize: 11,
-                  color: 'var(--color-text-quaternary)',
-                  flexShrink: 0,
-                  marginLeft: 8,
-                }}
-                title={relativeTime}
-              >
-                {relativeTime}
-              </span>
-            </div>
-          </div>
-          <div
-            className="todo-item-status"
-            aria-label="更改任务状态"
-          >
-            <StatusPicker
-              value={todo.status}
-              onChange={(newStatus) => handleStatusChange(todo.id, todo.title, todo.prompt || '', newStatus)}
-            />
-          </div>
-        </div>
-      </div>
-    );
-  };
-
-  if (isLoading) {
-    return (
-      <div className="todo-list-container">
-        <SkeletonList />
-      </div>
-    );
-  }
-
-  return (
-    <div className="todo-list-container">
-      {/* 搜索框：两种模式都展示（用户反馈：环路原本没有，切换时会跳界面）。
-          placeholder 按 listMode 切换，关键词同时匹配事项标题/提示词、环路名称。 */}
-      <div style={{ padding: '8px 16px', borderBottom: '1px solid var(--color-border-light)' }}>
-        <Input
-          placeholder={
-            listMode === 'item' ? '搜索标题或提示词...'
-            : '搜索环路名称...'
-          }
-          prefix={<SearchOutlined style={{ color: '#bfbfbf' }} />}
-          value={searchKeyword}
-          onChange={(e) => setSearchKeyword(e.target.value)}
-          allowClear
-          size="small"
-        />
-      </div>
-
-      {/* 通用操作工具栏：跨模式的「全选 / 批量 / 新建」入口。
-          createLabel / batchActions 按 listMode 在 toolbarConfig 中切换。 */}
-      <ActionToolbar
-        selectableIds={visibleIds}
-        selectedIds={selectedIds}
-        onSelectionChange={setSelectedIds}
-        createLabel={toolbarConfig.createLabel}
-        onCreate={
-          listMode === 'item' ? onOpenCreateModal
-          : onCreateLoop
-        }
-        batchActions={toolbarConfig.batchActions}
-        hideCreate={hideCreateButton}
-      />
-
-      {listMode === 'loop' ? (
-        <div style={{ flex: 1, minHeight: 0, overflow: 'auto' }}>
-          {loopLoading ? (
-            <Skeleton active style={{ padding: 16 }} />
-          ) : (
-            <LoopListPanel
-              loops={filteredLoopList}
-              selectedId={selectedLoopId}
-              onSelect={(id) => {
-                setSelectedLoopId(id);
-                onSelectLoop?.(id);
-              }}
-              onCreate={onCreateLoop}
-              selectedIds={selectedIds}
-              onToggleSelect={toggleSelect}
-              projectDirs={projectDirectories}
-              tags={tags}
-            />
-          )}
-        </div>
-      ) : (
-        <div className="todo-list-content">
-          {filteredTodos.length === 0 ? (
-            <div className="empty-state">
-              <div className="empty-state-icon">
-                <InboxOutlined />
-              </div>
-              <Empty
-                description={
-                  <div style={{ color: 'var(--color-text-tertiary)', fontSize: 14 }}>
-                    {selectedTagId ? '该标签下暂无任务' : '暂无任务'}
-                    <br />
-                    <span style={{ fontSize: 13, marginTop: 4, display: 'inline-block' }}>
-                      点击右上角新建按钮创建第一个任务
-                    </span>
-                  </div>
-                }
-                image={Empty.PRESENTED_IMAGE_SIMPLE}
-              />
-            </div>
-          ) : (
-            filteredTodos.map(renderTodoItem)
-          )}
-        </div>
-      )}
-
-      {/* 批量更换执行器 Modal：事项模式。
-          关闭即作废，不会触发回调（避免半路取消导致 selectedIds 与 Modal 状态不一致）。 */}
-      <Modal
-        title={`更换执行器（${pendingExecutorChangeIds.length} 项）`}
-        open={executorModalOpen}
-        onCancel={() => { setExecutorModalOpen(false); setPendingExecutorChangeIds([]); }}
-        footer={null}
-        destroyOnClose
-      >
-        <ExecutorPicker
-          executor=""
-          executorOptions={EXECUTORS_FOR_PICKER}
-          onChange={(v) => handleConfirmChangeExecutor(v)}
-        />
-      </Modal>
-
-      {/* 强停环路确认 Modal：占位实现，最终会调真实接口 */}
-      <Modal
-        title="强停环路"
-        open={forceStopModalOpen}
-        onOk={handleConfirmForceStop}
-        onCancel={() => { setForceStopModalOpen(false); setPendingForceStopIds([]); }}
-        okText="强停"
-        cancelText="取消"
-        okButtonProps={{ danger: true }}
-        destroyOnClose
-      >
-        <p>将停止 <strong>{pendingForceStopIds.length}</strong> 个环路关联的所有正在运行的执行。</p>
-        <p style={{ color: 'var(--color-text-tertiary)', fontSize: 12 }}>
-          （强停功能开发中，详见 utils/database/loops.ts 的 forceStopLoops 注释。）
-        </p>
-      </Modal>
-
-      {/* 批量工作空间操作 Modal（复制到/移动到） */}
-      <Modal
-        title={workspaceBatchMode === 'copy' ? '复制到工作空间' : '移动到工作空间'}
-        open={workspaceBatchModalOpen}
-        onOk={handleConfirmWorkspaceBatch}
-        onCancel={() => { setWorkspaceBatchModalOpen(false); setPendingWorkspaceBatchIds([]); }}
-        okText={workspaceBatchMode === 'copy' ? '确认复制' : '确认移动'}
-        cancelText="取消"
-        confirmLoading={workspaceBatchProcessing}
-        okButtonProps={{ disabled: workspaceBatchTarget == null }}
-        destroyOnClose
-      >
-        <p>
-          {workspaceBatchMode === 'copy' ? '复制' : '移动'} <strong>{pendingWorkspaceBatchIds.length}</strong> 项到目标工作空间：
-        </p>
-        <div style={{ marginTop: 12 }}>
-          <WorkspaceSelect
-            value={workspaceBatchTarget}
-            onChange={(v) => setWorkspaceBatchTarget(v ?? null)}
-            required
-          />
-        </div>
-        {workspaceBatchMode === 'copy' && (
-          <p style={{ color: 'var(--color-text-tertiary)', fontSize: 12, marginTop: 8 }}>
-            复制后，原工作空间和目标工作空间中各有一份相同的条目。
-          </p>
-        )}
-      </Modal>
-    </div>
-  );
-}
+export { TodoList, SkeletonRow, SkeletonList, TodoItemRow } from './todo-list';
